Extract frame metadata service
